hello

+++ BusManagement/src/main/java/com/example/dao/BusBookingRepository.java
@@ -4,6 +4,6 @@ import com.example.entites.BusBooking;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BusBookingRepository extends JpaRepository<BusBooking, Integer>{
+public interface BusBookingRepository extends JpaRepository<BusBooking, Long>{
     
 }
